refactor: 카카오 로그인 연동을 develop에 병합된 공용 API 함수 위로 재배선
develop에 이미 shared/api(access-token, api, auth-header 훅)가 있어
중복되던 shared/api/token.ts·entities/auth와 로컬 ADR(0004)을 걷어내고,
로그인 오케스트레이션이 그 공용 함수(setAccessToken/isAuthenticated)를
쓰도록 다시 연동했다.

- 401 세션 만료 처리(afterResponse)는 기존에 없던 부분이라 auth-header
  훅과 같은 자리(shared/api/hooks/session-expiry.ts)에 새로 추가
- 생성된 API 클라이언트는 apiClient.extend({ throwHttpErrors: false })로
  분리해, ky 기본 동작에 의존하는 shared/api의 다른 소비자(api.*)에는
  영향 없이 http-client.ts 자체 에러 처리와만 맞물리게 함
- 배포 도메인이 연동되어 VITE_API_BASE_URL과 generate:api 스펙 URL을
  https://api.aligneryoga.com 기준으로 전환하고 클라이언트 재생성
