+++ CONTEXT.md
@@ -12,5 +12,5 @@
 | ProgressRingItem | 원형 이미지 둘레를 진행 비율만큼 링으로 칠하고, 아래에 이름과 배지를 붙이는 아이템. `current=0`은 미시작(링·배지 없음), `current=total`은 완료 강조 | `shared/ui/progress-ring-item`. 단계 수는 `total`로 가변, 배지 문구는 외부 주입 |
 | SequenceItem | 좌측 순번 배지 + 점선 연결선, 우측 정보 카드로 이루어진 목록 한 줄. 순서가 있는 항목을 세로로 나열할 때 쓴다 | `shared/ui/sequence-item`. 마지막 줄은 `isLast`로 연결선을 숨긴다 |
 | Member | 카카오 로그인으로 식별·생성되는 회원 도메인 객체. 닉네임·키·몸무게·운동 경력 등 온보딩 정보를 갖는다. 온보딩 전에는 해당 필드가 null | `entities/member`. _Avoid_: User, 사용자, Account |
-| 세션 (Session) | 이 레포에서 "세션"은 항상 운동 세션(시작~완료 단위)을 가리킨다. 로그인/인증 여부를 가리킬 때는 "인증(Auth)"이라 부르고 "세션"이라 하지 않는다 | `entities/session`(운동, 추후)과 `entities/auth`(인증)로 분리해 이름 충돌을 피한다 |
+| 세션 (Session) | 이 레포에서 "세션"은 항상 운동 세션(시작~완료 단위)을 가리킨다. 로그인/인증 여부를 가리킬 때는 "인증(Auth)"이라 부르고 "세션"이라 하지 않는다 | `entities/session`(운동, 추후)과 `shared/api`의 인증 상태(`isAuthenticated`/`setAccessToken`)로 분리해 이름 충돌을 피한다 |
