--- src/entities/auth/index.ts
@@ -1,1 +1,0 @@
-export { isAuthenticated, login, logout } from "./model/auth";
